Use ES5 ponyfills for Array#find and String#endsWith.
Babel will not transpile prototypical methods even with plugins
like transform-runtime. We could leave it up to library users
to polyfill ES2015 methods, but that would be a burden. Since
we only use two such methods on a single line, we'll avoid the
issue by providing ponyfills. Use ponyfills instead of polyfills
to avoid conflict with other libraries.